Created super class for shot creation modals
